requirements.txt
streamlit
pandas
nltk
pillow
